Moved all changes from test.py into nexcsiserver.py; Added README for the branch
